requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
prophet>=1.1.0
lightgbm>=3.3.0
pyod>=1.0.0
adtk>=0.6.2
holidays>=0.14.0
openai>=1.0.0
python-dotenv>=1.0.0